Rewrite nthFibonacci tests for PhantomJS
Use `phantom` to load the page in a headless PhantomJS browser so
that the tests can access and evaluate the code that's loaded by
that page.

One concerning thing I've noticed about this approach is load time;
each test seems to take roughly 100ms, which is awfully long. The
approach may require further revision and refinement.